updating auto depth filtering to be compatible with vg call VCF formatting

+++ scripts/templates/04_vg_call_SVs.TEMPLATE.sh
@@ -25,12 +25,15 @@ INDIR="${GIR_OUT_DIR}"
 OUTDIR="${CALL_OUT_DIR}"
 
 SAMP="SAMP_NAME_k${KLEN}"
+SORTBAM="${SAMP}_subgraph.sorted.bam"
+DEDUPBAM="${SAMP}_subgraph.sorted.dedup.bam"
 SUBXG="mc_${TAXON}_all_chroms.gfa.${SAMP}.subgraph.gbz.xg"
 SORTGAM="${SAMP}_subgraph.sorted.gam"
 FILTGAM="${SAMP}_subgraph.sorted.filtered.gam"
 OUTVCF="${SAMP}_mc_subgraph.vcf"
 FILTVCF="${SAMP}_mc_subgraph_PASS.vcf"
 NORMVCF="${SAMP}_mc_subgraph_PASS.norm.vcf"
+SVVCF="${SAMP}_mc_subgraph_PASS.norm.SVsONLY.vcf.gz"
 SVVCF="${SAMP}_mc_subgraph_PASS.norm.SVsONLY.vcf.gz"
 
 LOGFILE="${LOG_DIR}/04_vg_call_${SAMP}_$(date +"%Y_%m_%d_%I_%M_%p").log"
@@ -93,31 +96,68 @@ fi
 # Normalize and filter to keep passing variants >= 50 bp in length
 # -----------------------------------------------------------------------------
 
-echo "(4/4) running VCF filtering - " $(date -u) >> ${LOGFILE}
-bcftools view \
-	-f PASS \
-	${OUTVCF} > \
-	${FILTVCF}
+if [ ! -f ${OUTDIR}/${SVVCF} ]; then
+	echo "(4/4) running VCF filtering - " $(date -u) >> ${LOGFILE}
+	bcftools view \
+		-f PASS \
+		${OUTVCF} > \
+		${FILTVCF}
 	
-bcftools norm \
-	-m-any \
-	-O z \
-	-o ${NORMVCF} \
-	${FILTVCF}
+	bcftools norm \
+		-m-any \
+		-O z \
+		-o ${NORMVCF} \
+		${FILTVCF}
 	
-grep "^#" ${NORMVCF} > tmp1.header1
-grep -v "^#CHROM" tmp1.header1 > tmp2.header1
-grep "^#CHROM" ${NORMVCF} > tmp.header2
-cat tmp2.header1 tmp.header2 > full.header
+	grep "^#" ${NORMVCF} > tmp1.header1
+	grep -v "^#CHROM" tmp1.header1 > tmp2.header1
+	grep "^#CHROM" ${NORMVCF} > tmp.header2
+	cat tmp2.header1 tmp.header2 > full.header
 
-grep -v "^#" ${NORMVCF} | \
-awk '{if (sqrt(((length($4)-length($5))^2))<50) print $0}' > \
-tmp.${SAMP}.vcf
+	grep -v "^#" ${NORMVCF} | \
+	awk '{if (sqrt(((length($4)-length($5))^2))<50) print $0}' > \
+	tmp.${SAMP}.vcf
 
-cat full.header tmp.${SAMP}.vcf | \
-gzip > \
-${SVVCF}
+	cat full.header tmp.${SAMP}.vcf | \
+	gzip > \
+	${SVVCF}
+else
+	rsync -avuP ${OUTDIR}/${SVVCF} .
+fi
 
+
+# -----------------------------------------------------------------------------
+# Calculate mean read depth and lower/upper depth limits for keeping variants
+# ! Based on giraffe BAM output, not GAM, so rough estimate ! 
+# -----------------------------------------------------------------------------
+
+if [ ! -f ${INDIR}/${DEDUPBAM} ]; then
+	rsync -avuP ${INDIR}/${SORTBAM} .
+	INBAM=${SORTBAM}
+else
+	rsync -avuP ${INDIR}/${DEDUPBAM} .
+	INBAM=${DEDUPBAM}
+fi
+
+samtools coverage ${INBAM} > tmp.coverage.txt
+
+AVG_DP=$(grep -v "^#" tmp.coverage.txt | awk '{ sum1+=($3*$7); } { sum2+=$3; } END{ print sum1/sum2;}')
+LO_LIM=$(awk -vdp=$AVG_DP -vlo=$LO_PROP 'BEGIN{printf "%.0f" ,dp * lo}')
+if (( $(echo "$LO_LIM $MIN" | awk '{print ($1 < $2)}') )); then
+	LO_LIM=$MIN
+fi
+UP_LIM=$(awk -vdp=$AVG_DP -vlo=$UP_PROP 'BEGIN{printf "%.0f" ,dp * lo}')
+
+DEPTHFILT_VCF="${SAMP}_mc_subgraph_PASS.norm.SVsONLY.d${LO_LIM}-${UP_LIM}.vcf.gz"
+	
+## filter to keep sites with acceptable read depths
+filt="(FMT/DP)>=${LO_LIM} & (FMT/DP)<=${UP_LIM}"
+bcftools filter -i "$filt" ${SVVCF} | bgzip -c ${VG_CALL_NTHREADS} > ${DEPTHFILT_VCF}
+
+# -----------------------------------------------------------------------------
+# Clean up tmp dir
+# -----------------------------------------------------------------------------
+rm tmp.coverage.txt
 rm tmp.${SAMP}.vcf
 rm ${FILTVCF}
 rm ${NORMVCF}
@@ -126,12 +166,6 @@ gzip *.vcf
 rsync -avuP *.vcf.gz ${OUTDIR}
 
 echo "SAMP_NAME - complete - " $(date -u) >> ${LOGFILE}
-	
-
-# -----------------------------------------------------------------------------
-# Clean up tmp dir
-# -----------------------------------------------------------------------------
-
 cd ${OUTDIR}
 rm -rf ${TMP_DIR}
